% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.r
\name{generator}
\alias{generator}
\title{Generate samples for multiple datasets}
\usage{
    generator(n, p, beta, distr = "gaussian", rho = 0.5)
}
\arguments{
\item{n}{The sample sizes in multiple datasets. An integer vector in \eqn{R^{M}}, where \eqn{M} is the number of datasets.}

\item{p}{The number of covariates. The number of covariates in each dataset is the same.}

\item{beta}{A \eqn{p \times M} numeric matrix, which is the true coefficients for the \eqn{M} datasets.}

\item{distr}{The error's distribution, including "gaussian", "t3", "cauchy" and "logistic".}

\item{rho}{The strength of correlation among covariates.}
}
\value{
A list
\itemize{
\item x - A \eqn{N \times p} design matrix, where \eqn{N} is the total sample size in \eqn{M} datasets.
\item y - A length \eqn{N} vector of the response for \eqn{M} datasets.
}
}
\description{
This function generates data including response and covariates for multiple datasets. 
Data can be heavy-tailed. 
Morever, there may be a high-correlated pattern among covariates.
}
\examples{
library("mnormt")
M = 2
n1 = n2 = 20
n = c(n1, n2)
p = 50
beta = matrix(0, p, M)
index1 = 1:10
index2 = 2:11
beta[index1, 1] = runif(10, 0.2, 1.0)
beta[index2, 2] = runif(10, 0.4, 1.4)
distr = "t3"
rho = 0.8
dat = generator(n, p, beta, distr, rho)
x = dat$x
y = dat$y
}
